Update Solution.java
Added problem statement and explained solution in comments.
Because there can be up to 10^6 cases, and each case can be up to 10^9, looping through all the values up to each number is not feasible. (Worst case ~10^15)
BigInteger was used because the maximum value of n is 10^9, causing the sum of all multiples of 3 and 5 to be easily in the range of ~10^17. (Much larger than 2^31-1)